Move some components
