+++ client/components/modals/DeleteFileWarning.js
@@ -1,9 +1,11 @@
 import React from 'react';
 
-import { useTranslation } from '../contexts/TranslationContext';
+import { useTranslation } from '../../contexts/TranslationContext';
 import DeleteWarningModal from './DeleteWarningModal';
 
-export default ({ onConfirm, onCancel }) => {
+const DeleteFileWarning = ({ onConfirm, onCancel }) => {
 	const t = useTranslation();
 	return <DeleteWarningModal onCancel={onCancel} onDelete={onConfirm}>{t('Delete_File_Warning')}</DeleteWarningModal>;
 };
+
+export default DeleteFileWarning;
