+++ client/components/modals/DeleteSuccessModal.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 import { Button, ButtonGroup, Icon, Modal } from '@rocket.chat/fuselage';
 
-import { useTranslation } from '../contexts/TranslationContext';
+import { useTranslation } from '../../contexts/TranslationContext';
 
 type DeleteSuccessModalProps = {
 	onClose: () => void;
