+++ client/views/admin/customUserStatus/EditCustomUserStatus.js
@@ -6,8 +6,8 @@ import { useMethod } from '../../../contexts/ServerContext';
 import { useToastMessageDispatch } from '../../../contexts/ToastMessagesContext';
 import { useSetModal } from '../../../contexts/ModalContext';
 import VerticalBar from '../../../components/VerticalBar';
-import DeleteSuccessModal from '../../../components/DeleteSuccessModal';
-import DeleteWarningModal from '../../../components/DeleteWarningModal';
+import DeleteSuccessModal from '../../../components/modals/DeleteSuccessModal';
+import DeleteWarningModal from '../../../components/modals/DeleteWarningModal';
 
 export function EditCustomUserStatus({ close, onChange, data, ...props }) {
 	const t = useTranslation();
