+++ client/views/admin/integrations/edit/EditIncomingWebhook.js
@@ -9,8 +9,8 @@ import { useToastMessageDispatch } from '../../../../contexts/ToastMessagesConte
 import { useSetModal } from '../../../../contexts/ModalContext';
 import { useForm } from '../../../../hooks/useForm';
 import IncomingWebhookForm from '../IncomingWebhookForm';
-import DeleteSuccessModal from '../../../../components/DeleteSuccessModal';
-import DeleteWarningModal from '../../../../components/DeleteWarningModal';
+import DeleteSuccessModal from '../../../../components/modals/DeleteSuccessModal';
+import DeleteWarningModal from '../../../../components/modals/DeleteWarningModal';
 import { useEndpointData } from '../../../../hooks/useEndpointData';
 import { AsyncStatePhase } from '../../../../hooks/useAsyncState';
