+++ client/views/admin/sidebar/AdminSidebar.js
@@ -1,7 +1,7 @@
 import React, { useCallback, useMemo, useEffect, memo } from 'react';
 
 import { menu, SideNav, Layout } from '../../../../app/ui-utils/client';
-import PlanTag from '../../../components/PlanTag';
+import PlanTag from './PlanTag';
 import Sidebar from '../../../components/Sidebar';
 import { useAtLeastOnePermission } from '../../../contexts/AuthorizationContext';
 import { useRoutePath, useCurrentRoute } from '../../../contexts/RouterContext';
