+++ ee/client/omnichannel/BusinessHoursTable.js
@@ -2,7 +2,7 @@ import { Table, Callout, Icon, Button } from '@rocket.chat/fuselage';
 import { useMutableCallback } from '@rocket.chat/fuselage-hooks';
 import React, { memo, useMemo, useState } from 'react';
 
-import DeleteWarningModal from '../../../client/components/DeleteWarningModal';
+import DeleteWarningModal from '../../../client/components/modals/DeleteWarningModal';
 import GenericTable from '../../../client/components/GenericTable';
 import FilterByText from '../../../client/components/FilterByText';
 import { useSetModal } from '../../../client/contexts/ModalContext';
